Ensure scripts dir is linked before inital service start
No ordering on the scripts File resource lets it sometimes drift to after the service start. The service appears to be creating this directory on startup and then the File resource fails to link because it's already a directory. This adds a `before` attribute matching other manged files/directories forcing it to happen before the service. Fixes #611